// server/types.ts
export interface BSDPluginSetup {
  config$: Observable<BSDConfig>;
}

export interface BSDPluginSetupDependencies {
  // required dependencies
  apmDataAccess: ApmDataAccessPluginSetup;
  data: DataPluginSetup;
  features: FeaturesPluginSetup;
  licensing: LicensingPluginSetup;
  observability: ObservabilityPluginSetup;
  ruleRegistry: RuleRegistryPluginSetupContract;
  metricsDataAccess: MetricsDataPluginSetup;
  dataViews: {};
  share: SharePluginSetup;
  observabilityAIAssistant?: ObservabilityAIAssistantServerSetup;
  // optional dependencies
  actions?: ActionsPlugin['setup'];
  alerting?: AlertingPlugin['setup'];
  cloud?: CloudSetup;
  fleet?: FleetPluginSetup;
  home?: HomeServerPluginSetup;
  ml?: MlPluginSetup;
  security?: SecurityPluginSetup;
  spaces?: SpacesPluginSetup;
  taskManager?: TaskManagerSetupContract;
  usageCollection?: UsageCollectionSetup;
  customIntegrations?: CustomIntegrationsPluginSetup;
  profilingDataAccess?: ProfilingDataAccessPluginSetup;
}
export interface BSDPluginStartDependencies {
  // required dependencies
  apmDataAccess: ApmDataAccessPluginStart;
  data: DataPluginStart;
  features: FeaturesPluginStart;
  licensing: LicensingPluginStart;
  observability: undefined;
  ruleRegistry: RuleRegistryPluginStartContract;
  metricsDataAccess: MetricsDataPluginSetup;
  dataViews: DataViewsServerPluginStart;
  share: undefined;
  observabilityAIAssistant?: ObservabilityAIAssistantServerStart;
  // optional dependencies
  actions?: ActionsPlugin['start'];
  alerting?: AlertingPlugin['start'];
  cloud?: undefined;
  fleet?: FleetPluginStart;
  home?: HomeServerPluginStart;
  ml?: MlPluginStart;
  security?: SecurityPluginStart;
  spaces?: SpacesPluginStart;
  taskManager?: TaskManagerStartContract;
  usageCollection?: undefined;
  customIntegrations?: CustomIntegrationsPluginStart;
  profilingDataAccess?: ProfilingDataAccessPluginStart;
}